+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 Changed `AccumulatorException.toString` to be simpler.
 Added warning if multiple registrations are present in `OAuth2Properties` when creating an `OAuth2TokenProvider`.
 Added `OAuth2TokenProvider` constructor without client registration name.
+Added `Temporals.formatToSeconds` to format a `double` to seconds as `String`.
+Changed `DurationAccumulator` to be faster and consume less memory.
 
 ---